backend/requirements.txt
flask==3.1.*
flask-cors==5.*
flask-sqlalchemy==3.1.*
flask-migrate==4.*
flask-jwt-extended==4.*
bcrypt==4.*
psycopg2-binary==2.9.*
python-dotenv==1.*
gunicorn==23.*
requests==2.32.*
beautifulsoup4==4.12.*
python-jobspy==1.*
linkedin-jobs-scraper==4.*
selenium==4.*
pytest==8.*
